Reorganize files and make conversion explicit

+++ Source/Constants/SpreadConstants.h
@@ -1,7 +1,6 @@
 #pragma once
 #include <cmath>
-#include "WorldConstants.h"
 
 const float SPREAD_DIST = 2.5f;
 const uint16_t MAX_SPREAD = 1024;
-const uint16_t SPREAD_PER_CHUNK = sqrt(MAX_SPREAD);
+const uint16_t SPREAD_PER_CHUNK = (uint16_t)sqrt(MAX_SPREAD);
